requirements.txt
requests>=2.26.0
python-dotenv>=0.19.1
loguru>=0.5.3
selenium>=4.1.0
undetected_chromedriver>=3.1.3
